gini-knowledge-engine-ecompai/requirements.txt
streamlit==1.34.0
fastapi
boto3
openai
python-multipart
pandas
mysql.connector
uvicorn
